refactor(sra): isolate pure views and canonical serialization (#171)
