+++ skills/sra/scripts/sra_serialization.py
@@ -1,0 +1,13 @@
+"""Canonical SRA serialization shared by packets, comparisons and runtime anchors."""
+from __future__ import annotations
+import hashlib
+import json
+from typing import Any
+
+
+def canonical_json(data: Any) -> str:
+    return json.dumps(data, ensure_ascii=False, sort_keys=True, separators=(",", ":"))
+
+
+def digest_data(data: Any) -> str:
+    return "sha256:" + hashlib.sha256(canonical_json(data).encode("utf-8")).hexdigest()
